requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
watchdog==3.0.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
synology-api==0.5.1
python-dotenv==1.0.0
pydantic<2.0
aiofiles==23.2.0
httpx==0.25.2
sqlalchemy==2.0.41
alembic==1.12.1
python-multipart==0.0.6
starlette==0.46.2
langchain==0.1.0
langchain-openai==0.0.2
openai==1.12.0
pinecone-client==3.0.0
tiktoken==0.5.2
celery==5.3.4
redis==5.0.1
prometheus-client==0.19.0
structlog==23.2.0